+++ .github/deploy/src/helpers/helper.ts
@@ -10,3 +10,13 @@ export const getSubDomainName = (props: GetSubDomainNameProps): string => {
   const { branchName } = props
   return isMasterBranch(branchName) ? '' : branchName
 }
+
+export interface GetUrlProps {
+  domainName: string;
+  subDomainName: string;
+}
+
+export const getUrl = (props: GetUrlProps): string => {
+  const { domainName, subDomainName } = props
+  return subDomainName === '' ? `${subDomainName}.${domainName}` : domainName
+}
