+++ .github/deploy/src/helpers/route53.ts
@@ -32,13 +32,3 @@ export const createARecordForDistribution = (props: CreateARecordForDistribution
     recordName: subDomainName
   })
 }
-
-export interface GetUrlProps {
-  domainName: string;
-  subDomainName: string;
-}
-
-export const getUrl = (props: GetUrlProps): string => {
-  const { domainName, subDomainName } = props
-  return subDomainName === '' ? `${subDomainName}.${domainName}` : domainName
-}
